Add files via upload
This is a command line program that is used as a simple database to save student information. To everyone that needs it as a course assignment or to use it as a core to develop a more sophisticated database this is a pretty straight forward one-file project less than 500 lines.